generators/requirements.txt
# generators/requirements.txt
kafka-python==2.0.2
